refactor(mcp): extract the MCP manager into internal/mcp
Move-only: relocate mcp_manager.go and mcp_sse_transport.go (+ their tests)
into a new internal/mcp context and drop the now-redundant mcp_ filename
prefix. The manager implements agentdomain.MCPManager and MCPClient; the SSE
transport has zero internal imports and only ever served the manager.

Only the package clause and container.go's import change.
