+++ internal/mcp/manager.go
@@ -1,4 +1,4 @@
-package services
+package mcp
 
 import (
 	"context"
