+++ internal/mcp/sse_transport.go
@@ -1,4 +1,4 @@
-package services
+package mcp
 
 import (
 	"bufio"
