+++ internal/mcp/sse_transport_test.go
@@ -1,4 +1,4 @@
-package services
+package mcp
 
 import (
 	"context"
